MAINT: move all SWT-related python functions into _swt.py
