+++ pywt/_multilevel.py
@@ -20,7 +20,7 @@ from ._dwt import dwt, idwt
 from ._multidim import dwt2, idwt2, dwtn, idwtn, _fix_coeffs
 
 __all__ = ['wavedec', 'waverec', 'wavedec2', 'waverec2', 'wavedecn',
-           'waverecn', 'iswt', 'iswt2', 'coeffs_to_array', 'array_to_coeffs']
+           'waverecn', 'coeffs_to_array', 'array_to_coeffs']
 
 
 def _check_level(size, dec_len, level):
@@ -256,181 +256,6 @@ def waverec2(coeffs, wavelet, mode='symmetric'):
         a = idwt2((a[idxs], d), wavelet, mode)
 
     return a
-
-
-def iswt(coeffs, wavelet):
-    """
-    Multilevel 1D Inverse Discrete Stationary Wavelet Transform.
-
-    Parameters
-    ----------
-    coeffs : array_like
-        Coefficients list of tuples::
-
-            [(cAn, cDn), ..., (cA2, cD2), (cA1, cD1)]
-
-        where cA is approximation, cD is details.  Index 1 corresponds to
-        ``start_level`` from ``pywt.swt``.
-    wavelet : Wavelet object or name string
-        Wavelet to use
-
-    Returns
-    -------
-    1D array of reconstructed data.
-
-    Examples
-    --------
-    >>> import pywt
-    >>> coeffs = pywt.swt([1,2,3,4,5,6,7,8], 'db2', level=2)
-    >>> pywt.iswt(coeffs, 'db2')
-    array([ 1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.])
-    """
-
-    output = coeffs[0][0].copy()  # Avoid modification of input data
-
-    # num_levels, equivalent to the decomposition level, n
-    num_levels = len(coeffs)
-    if not isinstance(wavelet, Wavelet):
-        wavelet = Wavelet(wavelet)
-    for j in range(num_levels, 0, -1):
-        step_size = int(pow(2, j-1))
-        last_index = step_size
-        _, cD = coeffs[num_levels - j]
-        for first in range(last_index):  # 0 to last_index - 1
-
-            # Getting the indices that we will transform
-            indices = np.arange(first, len(cD), step_size)
-
-            # select the even indices
-            even_indices = indices[0::2]
-            # select the odd indices
-            odd_indices = indices[1::2]
-
-            # perform the inverse dwt on the selected indices,
-            # making sure to use periodic boundary conditions
-            x1 = idwt(output[even_indices], cD[even_indices],
-                      wavelet, 'periodization')
-            x2 = idwt(output[odd_indices], cD[odd_indices],
-                      wavelet, 'periodization')
-
-            # perform a circular shift right
-            x2 = np.roll(x2, 1)
-
-            # average and insert into the correct indices
-            output[indices] = (x1 + x2)/2.
-
-    return output
-
-
-def iswt2(coeffs, wavelet):
-    """
-    Multilevel 2D Inverse Discrete Stationary Wavelet Transform.
-
-    Parameters
-    ----------
-    coeffs : list
-        Approximation and details coefficients::
-
-            [
-                (cA_1,
-                    (cH_1, cV_1, cD_1)
-                ),
-                (cA_2,
-                    (cH_2, cV_2, cD_2)
-                ),
-                ...,
-                (cA_n
-                    (cH_n, cV_n, cD_n)
-                )
-            ]
-
-        where cA is approximation, cH is horizontal details, cV is
-        vertical details, cD is diagonal details and n is the number of
-        levels.  Index 1 corresponds to ``start_level`` from ``pywt.swt2``.
-    wavelet : Wavelet object or name string
-        Wavelet to use
-
-    Returns
-    -------
-    2D array of reconstructed data.
-
-    Examples
-    --------
-    >>> import pywt
-    >>> coeffs = pywt.swt2([[1,2,3,4],[5,6,7,8],
-    ...                     [9,10,11,12],[13,14,15,16]],
-    ...                    'db1', level=2)
-    >>> pywt.iswt2(coeffs, 'db1')
-    array([[  1.,   2.,   3.,   4.],
-           [  5.,   6.,   7.,   8.],
-           [  9.,  10.,  11.,  12.],
-           [ 13.,  14.,  15.,  16.]])
-
-    """
-
-    output = coeffs[-1][0].copy()  # Avoid modification of input data
-
-    # num_levels, equivalent to the decomposition level, n
-    num_levels = len(coeffs)
-    if not isinstance(wavelet, Wavelet):
-        wavelet = Wavelet(wavelet)
-    for j in range(num_levels, 0, -1):
-        step_size = int(pow(2, j-1))
-        last_index = step_size
-        _, (cH, cV, cD) = coeffs[j-1]
-        # We are going to assume cH, cV, and cD are square and of equal size
-        if (cH.shape != cV.shape) or (cH.shape != cD.shape) or (
-                cH.shape[0] != cH.shape[1]):
-            raise RuntimeError(
-                "Mismatch in shape of intermediate coefficient arrays")
-        for first_h in range(last_index):  # 0 to last_index - 1
-            for first_w in range(last_index):  # 0 to last_index - 1
-                # Getting the indices that we will transform
-                indices_h = slice(first_h, cH.shape[0], step_size)
-                indices_w = slice(first_w, cH.shape[1], step_size)
-
-                even_idx_h = slice(first_h, cH.shape[0], 2*step_size)
-                even_idx_w = slice(first_w, cH.shape[1], 2*step_size)
-                odd_idx_h = slice(first_h + step_size, cH.shape[0], 2*step_size)
-                odd_idx_w = slice(first_w + step_size, cH.shape[1], 2*step_size)
-
-                # perform the inverse dwt on the selected indices,
-                # making sure to use periodic boundary conditions
-                x1 = idwt2((output[even_idx_h, even_idx_w],
-                           (cH[even_idx_h, even_idx_w],
-                            cV[even_idx_h, even_idx_w],
-                            cD[even_idx_h, even_idx_w])),
-                           wavelet, 'periodization')
-                x2 = idwt2((output[even_idx_h, odd_idx_w],
-                           (cH[even_idx_h, odd_idx_w],
-                            cV[even_idx_h, odd_idx_w],
-                            cD[even_idx_h, odd_idx_w])),
-                           wavelet, 'periodization')
-                x3 = idwt2((output[odd_idx_h, even_idx_w],
-                           (cH[odd_idx_h, even_idx_w],
-                            cV[odd_idx_h, even_idx_w],
-                            cD[odd_idx_h, even_idx_w])),
-                           wavelet, 'periodization')
-                x4 = idwt2((output[odd_idx_h, odd_idx_w],
-                           (cH[odd_idx_h, odd_idx_w],
-                            cV[odd_idx_h, odd_idx_w],
-                            cD[odd_idx_h, odd_idx_w])),
-                           wavelet, 'periodization')
-
-                # perform a circular shifts
-                x2 = np.roll(x2, 1, axis=1)
-                x3 = np.roll(x3, 1, axis=0)
-                x4 = np.roll(x4, 1, axis=0)
-                x4 = np.roll(x4, 1, axis=1)
-                output[indices_h, indices_w] = (x1 + x2 + x3 + x4) / 4
-
-    warnings.warn(
-        "For consistency with the rest of PyWavelets, the order of levels in "
-        "coeffs expected by iswt2 will be reversed in the next release. "
-        "In other words, the levels will be sorted in descending rather than "
-        "ascending order.", FutureWarning)
-
-    return output
 
 
 def wavedecn(data, wavelet, mode='symmetric', level=None):
